Unify telemetry and add Gleap widget for feedback
- Replace PostHog with unified telemetry system
- Add Gleap widget integration for user feedback
- Remove legacy feedback modal and database schema
- Update environment configuration for Gleap API key
- Consolidate telemetry providers and utilities
