+++ apps/web/client/.env.example
@@ -25,6 +25,9 @@ RELACE_API_KEY="<Your api key from https://app.relace.ai/settings/api-keys>"
 NEXT_PUBLIC_POSTHOG_KEY="<Your PostHog API key from https://posthog.com/docs/libraries/next-js>"
 NEXT_PUBLIC_POSTHOG_HOST="<Your PostHog region https://us.i.posthog.com>"
 
+# Gleap - In-app feedback & bug reports
+NEXT_PUBLIC_GLEAP_API_KEY="<Gleap support API Key>"
+
 # RB2B - Enables B2B visitor identification
 NEXT_PUBLIC_RB2B_ID="<Your RB2B tracking script ID from https://app.rb2b.com>"
 
@@ -53,10 +56,6 @@ N8N_API_KEY="<Your n8n api key>"
 LANGFUSE_SECRET_KEY=
 LANGFUSE_PUBLIC_KEY=
 LANGFUSE_BASEURL=
-
-# Feedback - Email notifications for user feedback
-FEEDBACK_FROM_EMAIL="contact@onlook.com"
-FEEDBACK_TO_EMAIL="support@onlook.com"
 
 # ------------- Optional: Alternative LLM providers -------------
