+++ apps/web/client/package.json
@@ -74,6 +74,7 @@
         "date-fns": "^4.1.0",
         "flexsearch": "^0.8.160",
         "freestyle-sandboxes": "^0.0.78",
+        "gleap": "^14.8.8",
         "langfuse-vercel": "^3.38.4",
         "localforage": "^1.10.0",
         "lru-cache": "^11.2.1",
