+++ apps/web/client/src/app/features/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { CreateManagerProvider } from '@/components/store/create';
-import { FeedbackModal } from '@/components/ui/feedback-modal';
 import { SubscriptionModal } from '@/components/ui/pricing-modal';
 import { NonProjectSettingsModal } from '@/components/ui/settings-modal/non-project';
 import { FeaturesHero } from '../_components/hero/features-hero';
@@ -58,7 +57,6 @@ export default function FeaturesPage() {
                 <FAQSection faqs={featuresFaqs} />
                 <NonProjectSettingsModal />
                 <SubscriptionModal />
-                <FeedbackModal />
             </WebsiteLayout>
         </CreateManagerProvider>
     );
