+++ apps/web/client/src/app/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { CreateManagerProvider } from '@/components/store/create';
-import { FeedbackModal } from '@/components/ui/feedback-modal';
 import { SubscriptionModal } from '@/components/ui/pricing-modal';
 import { NonProjectSettingsModal } from '@/components/ui/settings-modal/non-project';
 import { AuthModal } from './_components/auth-modal';
@@ -9,10 +8,10 @@ import { Hero } from './_components/hero';
 import { ContributorSection } from './_components/landing-page/contributor-section';
 import { CTASection } from './_components/landing-page/cta-section';
 import { FAQSection } from './_components/landing-page/faq-section';
+import { ResponsiveMockupSection } from './_components/landing-page/responsive-mockup-section';
 import { TestimonialsSection } from './_components/landing-page/testimonials-section';
 import { WhatCanOnlookDoSection } from './_components/landing-page/what-can-onlook-do-section';
 import { WebsiteLayout } from './_components/website-layout';
-import { ResponsiveMockupSection } from './_components/landing-page/responsive-mockup-section';
 
 export default function Main() {
     return (
@@ -22,17 +21,16 @@ export default function Main() {
                     <Hero />
                 </div>
                 <ResponsiveMockupSection />
-                {/* <CodeOneToOneSection /> */}                
+                {/* <CodeOneToOneSection /> */}
                 <WhatCanOnlookDoSection />
                 {/* <ObsessForHoursSection /> */}
                 <ContributorSection />
-                <TestimonialsSection />                
+                <TestimonialsSection />
                 <FAQSection />
                 <CTASection />
                 <AuthModal />
                 <NonProjectSettingsModal />
                 <SubscriptionModal />
-                <FeedbackModal />
             </WebsiteLayout >
         </CreateManagerProvider>
     );
