+++ apps/web/client/src/app/project/[id]/_components/main.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { FeedbackModal } from '@/components/ui/feedback-modal';
 import { SubscriptionModal } from '@/components/ui/pricing-modal';
 import { SettingsModalWithProjects } from '@/components/ui/settings-modal/with-project';
 import { EditorAttributes } from '@onlook/constants';
@@ -123,7 +122,6 @@ export const Main = observer(() => {
             </div>
             <SettingsModalWithProjects />
             <SubscriptionModal />
-            <FeedbackModal />
         </TooltipProvider>
     );
 });
