+++ apps/web/client/src/components/store/state/manager.ts
@@ -4,7 +4,6 @@ import { makeAutoObservable } from 'mobx';
 export class StateManager {
     isSubscriptionModalOpen = false;
     isSettingsModalOpen = false;
-    isFeedbackModalOpen = false;
     settingsTab: SettingsTabValue | string = SettingsTabValue.SITE;
 
     constructor() {
