+++ packages/email/src/index.ts
@@ -1,4 +1,3 @@
 export * from './client';
-export * from './feedback';
 export * from './invitation';
 export * from './templates';
