+++ packages/email/src/templates/index.ts
@@ -1,2 +1,1 @@
-export * from './feedback-notification';
 export * from './invite-user';
